create CxPropertiesBase
