+++ src/main/java/com/checkmarx/sdk/config/CxProperties.java
@@ -11,44 +11,31 @@ import javax.annotation.PostConstruct;
 @Component
 @ConfigurationProperties(prefix = "checkmarx")
 @Validated
-public class CxProperties {
+public class CxProperties extends CxPropertiesBase{
     public static final String CONFIG_PREFIX = "sast";
-
-    private Double version = 8.9;
+    
     private String username;
     private String password;
     private String clientId = "resource_owner_client";
     private String soapClientId = "resource_owner_sast_client";
-    private String clientSecret;
     private String soapClientSecret;
     private String scope = "sast_rest_api";
     private String soapScope = "sast_api";
-    private String baseUrl;
-    private String url;
     private String acUrl;
-    private boolean multiTenant = true;
     private String teamScript;
     private String projectScript;
-    private String scanPreset = Constants.CX_DEFAULT_PRESET;
-    private String configuration = Constants.CX_DEFAULT_CONFIGURATION;
     private String excludeFiles;
     private String excludeFolders;
     private Boolean incremental = false;
     private Boolean enableOsa = false;
-    private String gitClonePath;
     private Integer incrementalThreshold = 7;
     private Integer incrementalNumScans = 5;
-    private String team;
     private Boolean offline = false;
     private Boolean preserveXml = false;
-    private Integer scanTimeout = 120;
-    private String jiraProjectField = "jira-project";
-    private String jiraIssuetypeField = "jira-issuetype";
-    private String jiraCustomField = "jira-fields";
-    private String jiraAssigneeField = "jira-assignee";
+    
     private Integer httpConnectionTimeout = 30000;
     private Integer httpReadTimeout = 120000;
-    private Integer scanPolling = 20000;
+
     private Integer reportPolling = 5000;
     private Integer reportTimeout = 300000;
     private Integer codeSnippetLength = 2500;
@@ -57,8 +44,6 @@ public class CxProperties {
     private Boolean enableShardManager = false;
     private Boolean enablePostActionMonitor = false;
     private Integer postActionPostbackId = 0;
-
-    private String portalUrl;
 
     private String portalPackage = "checkmarx.wsdl.portal";
 
@@ -87,22 +72,7 @@ public class CxProperties {
     public String getClientId() {
         return this.clientId;
     }
-
-    public String getClientSecret() {
-        return this.clientSecret;
-    }
-
-    public String getBaseUrl() {
-        return this.baseUrl;
-    }
-
-    public String getUrl() {
-        return this.url;
-    }
-
-    public boolean isMultiTenant() {
-        return this.multiTenant;
-    }
+    
 
     public Boolean getIncremental() {
         return this.incremental;
@@ -114,18 +84,6 @@ public class CxProperties {
 
     public Integer getIncrementalNumScans(){
         return this.incrementalNumScans;
-    }
-
-    public String getScanPreset() {
-        return this.scanPreset;
-    }
-
-    public String getConfiguration() {
-        return this.configuration;
-    }
-
-    public String getTeam() {
-        return this.team;
     }
 
     public Boolean getOffline() {
@@ -140,26 +98,7 @@ public class CxProperties {
         this.preserveXml = preserveXml;
     }
 
-    public Integer getScanTimeout() {
-        return this.scanTimeout;
-    }
-
-    public String getJiraProjectField() {
-        return this.jiraProjectField;
-    }
-
-    public String getJiraCustomField() {
-        return this.jiraCustomField;
-    }
-
-    public String getJiraIssuetypeField() {
-        return this.jiraIssuetypeField;
-    }
-
-    public String getPortalUrl() {
-        return this.portalUrl;
-    }
-
+ 
     public String getPortalPackage() {
         return this.portalPackage;
     }
@@ -179,23 +118,7 @@ public class CxProperties {
     public void setClientId(String clientId) {
         this.clientId = clientId;
     }
-
-    public void setClientSecret(String clientSecret) {
-        this.clientSecret = clientSecret;
-    }
-
-    public void setBaseUrl(String baseUrl) {
-        this.baseUrl = baseUrl;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public void setMultiTenant(boolean multiTenant) {
-        this.multiTenant = multiTenant;
-    }
-
+    
     public String getTeamScript() {
         return teamScript;
     }
@@ -236,42 +159,12 @@ public class CxProperties {
         this.incrementalNumScans = incrementalNumScans;
     }
 
-    public void setScanPreset(String scanPreset) {
-        this.scanPreset = scanPreset;
-    }
-
-    public void setConfiguration(String configuration) {
-        this.configuration = configuration;
-    }
-
-    public void setTeam(String team) {
-        this.team = team;
-    }
-
+ 
     public void setOffline(Boolean offline) {
         this.offline = offline;
     }
-
-    public void setScanTimeout(Integer scanTimeout) {
-        this.scanTimeout = scanTimeout;
-    }
-
-    public void setJiraProjectField(String jiraProjectField) {
-        this.jiraProjectField = jiraProjectField;
-    }
-
-    public void setJiraIssuetypeField(String jiraIssuetypeField) {
-        this.jiraIssuetypeField = jiraIssuetypeField;
-    }
-
-    public void setJiraCustomField(String jiraCustomField) {
-        this.jiraCustomField = jiraCustomField;
-    }
-
-    public void setPortalUrl(String portalUrl) {
-        this.portalUrl = portalUrl;
-    }
-
+    
+  
     public void setPortalPackage(String portalPackage) {
         this.portalPackage = portalPackage;
     }
@@ -280,14 +173,7 @@ public class CxProperties {
         this.htmlStrip = htmlStrip;
     }
 
-    public String getJiraAssigneeField() {
-        return this.jiraAssigneeField;
-    }
-
-    public void setJiraAssigneeField(String jiraAssigneeField) {
-        this.jiraAssigneeField = jiraAssigneeField;
-    }
-
+ 
     public String getExcludeFiles() {
         return excludeFiles;
     }
@@ -320,14 +206,6 @@ public class CxProperties {
         this.httpReadTimeout = httpReadTimeout;
     }
 
-    public Integer getScanPolling() {
-        return scanPolling;
-    }
-
-    public void setScanPolling(Integer scanPolling) {
-        this.scanPolling = scanPolling;
-    }
-
     public Integer getReportTimeout() {
         return reportTimeout;
     }
@@ -335,15 +213,7 @@ public class CxProperties {
     public void setReportTimeout(Integer reportTimeout) {
         this.reportTimeout = reportTimeout;
     }
-
-    public Double getVersion() {
-        return version;
-    }
-
-    public void setVersion(Double version) {
-        this.version = version;
-    }
-
+    
     public String getAcUrl() {
         return acUrl;
     }
@@ -411,40 +281,6 @@ public class CxProperties {
     public void setSoapScope(String soapScope) {
         this.soapScope = soapScope;
     }
-
-    public String getTeamPathSeparator(){
-        if(version < 9.0){
-            return TEAM_PATH_SEPARATOR_8;
-        }
-        else{
-            return TEAM_PATH_SEPARATOR_9;
-        }
-    }
-
-    public String getGitClonePath(){
-        if(this.gitClonePath == null){
-            if (System.getProperty("os.name").startsWith("Windows")) {
-                // includes: Windows 2000,  Windows 95, Windows 98, Windows NT, Windows Vista, Windows XP
-                return Constants.WINDOWS_PATH;
-            } else {
-                // everything else
-                return Constants.UNIX_PATH;
-            }
-        }
-        else {
-            return this.gitClonePath;
-        }
-    }
-
-    public void setGitClonePath(String gitClonePath) {
-        this.gitClonePath = gitClonePath;
-    }
-
-    @PostConstruct
-    private void initTeam(){
-        if(team != null && !team.startsWith(getTeamPathSeparator())){
-            this.team = getTeamPathSeparator().concat(this.team);
-        }
-    }
+    
 }
